Перенёс интерфейсы репозиториев в отдельный модуль
