+++ data/build.gradle
@@ -9,4 +9,6 @@ java {
 
 dependencies {
     api(project(":model"))
+
+    api "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines_version"
 }
